projects/hackathon-contracts/smart_contracts/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "email-validator>=2.1",
    "fastapi>=0.119.0",
    "firebase-admin>=7.1.0",
    "google-auth>=2.41.1",
    "google-cloud-firestore>=2.18",
    "itsdangerous>=2.2.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn[standard]>=0.37.0",
]
